refactor : organisms
