+++ src/components/organisms/AlertModal/index.js
@@ -1,0 +1,1 @@
+export { AlertModal } from "./AlertModal";
