+++ src/components/organisms/LoadingModal/index.js
@@ -1,0 +1,1 @@
+export { LoadingModal } from "./LoadingModal";
