--- src/components/organisms/LoadingModal/style.jsx
@@ -1,15 +1,0 @@
-import styled from "styled-components";
-
-export const Background = styled.div`
-  height: 100%;
-  width: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  z-index: 999;
-  position: fixed;
-  left: 0;
-  top: 0;
-  text-align: center;
-  background-color: rgba(0, 0, 0, 0.1);
-`;
