+++ src/components/organisms/MySection/index.js
@@ -1,0 +1,1 @@
+export { MySection } from "./MySection";
