--- src/components/organisms/MySection/style.jsx
@@ -1,55 +1,0 @@
-import styled from "styled-components";
-
-export const MyReceiptContainer = styled.section`
-  box-sizing: border-box;
-  position: relative;
-  width: 100%;
-  padding: 2.75rem 1rem;
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  justify-content: center;
-`;
-
-export const HeaderContainer = styled.header`
-  margin-bottom: 1rem;
-`;
-
-export const MainText = styled.p`
-  margin-bottom: 0.5rem;
-  font-size: 1.5rem;
-  font-weight: bold;
-`;
-export const Box = styled.div`
-  display: flex-shrink;
-`;
-
-export const MyReceiptSeeAll = styled.div`
-  display: flex;
-  width: 20rem;
-  overflow-y: hidden;
-  overflow-x: scroll;
-  margin-bottom: 1rem;
-`;
-
-export const ReceiptImg = styled.img`
-  margin-right: 1rem;
-  width: 20rem;
-  height: 100%;
-`;
-export const ButtonContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 100%;
-`;
-
-export const Button = styled.button`
-  text-align: center;
-  width: 300px;
-  height: 80px;
-  text-align: center;
-  width: 300px;
-  height: 80px;
-  margin-right: 11.8rem;
-`;
